+++ tests/test_database.py
@@ -1,4 +1,6 @@
 """
+Tests related to the databases.
+
 `# type: ignore` are mostly do to mypy's too strict way of handling the type of property getters and setters.
 More details here: https://github.com/python/mypy/issues/3004
 """